Changed the participant payout to server side for faster loading.
