+++ src/app/(pages)/pool/[pool-id]/participants/[participant-id]/_components/payout-form.tsx
@@ -1,0 +1,111 @@
+'use client'
+
+import { useEffect, useRef, useState } from 'react'
+import { usePayoutStore } from '@/app/_client/stores/payout-store'
+
+import { Button } from '@/app/_components/ui/button'
+import { Input } from '@/app/_components/ui/input'
+import { cn } from '@/lib/utils/tailwind'
+
+import { toast } from 'sonner'
+import { Address, formatUnits, getAbiItem, parseUnits } from 'viem'
+import { useWriteContract } from 'wagmi'
+import { useTokenDecimals } from '@/app/(pages)/profile/send/_components/use-token-decimals'
+import { poolAbi } from '@/types/contracts'
+import useTransactions from '@/app/_client/hooks/use-transactions'
+import { currentPoolAddress } from '@/app/_server/blockchain/server-config'
+
+interface PayoutFormProps {
+    poolId: string
+    participantId: Address
+    tokenAddress: Address
+}
+
+const PayoutForm: React.FC<PayoutFormProps> = ({ poolId, participantId, tokenAddress }) => {
+    const { tokenDecimalsData } = useTokenDecimals(tokenAddress)
+    const { data: hash, isPending, isSuccess } = useWriteContract()
+    const { executeTransactions } = useTransactions()
+    const { addPayout, getPayoutForParticipant } = usePayoutStore()
+
+    const inputRef = useRef<HTMLInputElement | null>(null)
+    const [inputValue, setInputValue] = useState<string>('0')
+
+    const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+        setInputValue(event.target.value)
+    }
+
+    const onSaveButtonClicked = () => {
+        const amount = parseUnits(inputValue, tokenDecimalsData?.tokenDecimals)
+        addPayout(poolId.toString(), { amount: amount.toString(), participantAddress: participantId })
+        toast.success('Payout saved successfully')
+    }
+
+    const onPayoutButtonClicked = () => {
+        const SetWinnerFunction = getAbiItem({
+            abi: poolAbi,
+            name: 'setWinner',
+        })
+        const winnerAmount = parseUnits(inputValue, tokenDecimalsData?.tokenDecimals)
+
+        const args = [
+            {
+                address: currentPoolAddress,
+                abi: [SetWinnerFunction],
+                functionName: SetWinnerFunction.name,
+                args: [BigInt(poolId), participantId, winnerAmount],
+            },
+        ]
+
+        try {
+            executeTransactions(args)
+        } catch (error) {
+            console.log('setWinner Error', error)
+        }
+    }
+
+    useEffect(() => {
+        const savedPayout = getPayoutForParticipant(poolId.toString(), participantId)
+        if (savedPayout) {
+            setInputValue(formatUnits(BigInt(savedPayout.amount), tokenDecimalsData?.tokenDecimals).toString())
+        }
+    }, [poolId, participantId, getPayoutForParticipant, tokenDecimalsData?.tokenDecimals])
+
+    useEffect(() => {
+        if (isSuccess) {
+            toast.success('Payout Successful', { description: `Transaction: ${hash}` })
+        }
+    }, [isPending, hash, isSuccess])
+
+    return (
+        <>
+            <div className='mt-2 flex h-16 flex-row justify-center'>
+                <div className='relative flex justify-center'>
+                    <Input
+                        className={cn('h-24 w-auto border-none text-center text-6xl font-bold focus:outline-none')}
+                        placeholder='$'
+                        autoFocus={true}
+                        value={inputValue}
+                        type='number'
+                        onChange={handleInputChange}
+                        ref={inputRef}
+                        inputMode='numeric'
+                    />
+                </div>
+            </div>
+            <div className='mt-8 flex w-full flex-col items-center justify-center space-y-2'>
+                <Button
+                    onClick={onPayoutButtonClicked}
+                    className='mb-3 h-[46px] w-full flex-1 grow flex-row items-center justify-center rounded-[2rem] bg-cta py-[11px] text-center align-middle font-semibold leading-normal text-white shadow-button active:shadow-button-push'>
+                    Payout
+                </Button>
+                <Button
+                    onClick={onSaveButtonClicked}
+                    className='mb-3 h-[46px] w-full flex-1 grow flex-row items-center justify-center rounded-[2rem] py-[11px] text-center align-middle font-semibold leading-normal text-white shadow-button active:shadow-button-push'>
+                    Save
+                </Button>
+            </div>
+        </>
+    )
+}
+
+export default PayoutForm
